+++ data.tf
@@ -60,54 +60,47 @@ data "cloudinit_config" "k3s_server" {
 
   part {
     content_type = "text/x-shellscript"
-    content = templatefile("${path.module}/files/k3s-install-server.sh", {
-      k3s_version                       = local.k3s_version
-      k3s_subnet                        = var.k3s_subnet
-      k3s_token                         = var.enable_vault ? "" : random_password.k3s_token.result
-      disable_ingress                   = var.disable_ingress
-      certmanager_email_address         = var.certmanager_email_address
-      compartment_ocid                  = var.compartment_ocid
-      availability_domain               = var.availability_domain
-      cluster_name                      = var.cluster_name
-      k3s_url                           = local.k3s_internal_lb_ip
-      k3s_tls_san                       = local.k3s_internal_lb_ip
-      expose_kubeapi                    = var.expose_kubeapi
-      k3s_tls_san_public                = try(local.public_lb_ip[0], "")
-      argocd_hostname                   = var.argocd_hostname != null ? var.argocd_hostname : ""
-      longhorn_hostname                 = var.longhorn_hostname != null ? var.longhorn_hostname : ""
-      longhorn_ui_username              = var.longhorn_ui_username
-      longhorn_ui_password              = var.enable_vault ? "" : random_password.longhorn_ui_password.result
-      grafana_admin_password            = var.enable_vault ? "" : random_password.grafana_admin_password.result
-      gitops_repo_url                   = var.gitops_repo_url
-      kured_start_time                  = var.kured_start_time
-      kured_end_time                    = var.kured_end_time
-      kured_reboot_days                 = join(",", var.kured_reboot_days)
-      ingress_controller_http_nodeport  = var.ingress_controller_http_nodeport
-      ingress_controller_https_nodeport = var.ingress_controller_https_nodeport
-      notification_topic_endpoint       = var.enable_notifications ? oci_ons_notification_topic.k3s_alerts[0].api_endpoint : ""
-      mysql_endpoint                    = var.enable_mysql ? "${oci_mysql_mysql_db_system.k3s[0].endpoints[0].hostname}:${oci_mysql_mysql_db_system.k3s[0].endpoints[0].port}" : ""
-      mysql_admin_username              = var.enable_mysql ? var.mysql_admin_username : ""
-      mysql_admin_password              = var.enable_mysql ? random_password.mysql_admin_password[0].result : ""
-      vault_secret_id_k3s_token         = var.enable_vault ? oci_vault_secret.k3s_token[0].id : ""
-      vault_secret_id_longhorn_password = var.enable_vault ? oci_vault_secret.longhorn_ui_password[0].id : ""
-      vault_secret_id_grafana_password  = var.enable_vault ? oci_vault_secret.grafana_admin_password[0].id : ""
-      gateway_api_version               = var.gateway_api_version
-      envoy_gateway_chart_version       = var.envoy_gateway_chart_version
-      certmanager_chart_version         = var.certmanager_chart_version
-      longhorn_chart_version            = var.longhorn_chart_version
-      argocd_chart_version              = var.argocd_chart_version
-      kured_chart_version               = var.kured_chart_version
-      enable_external_dns               = var.enable_external_dns
-      cloudflare_api_token              = var.cloudflare_api_token != null ? var.cloudflare_api_token : ""
-      cloudflare_zone_id                = var.cloudflare_zone_id != null ? var.cloudflare_zone_id : ""
-      external_dns_domain_filter        = var.external_dns_domain_filter != null ? var.external_dns_domain_filter : ""
-      external_dns_chart_version        = var.external_dns_chart_version
-      enable_external_secrets           = var.enable_external_secrets
-      vault_ocid                        = var.enable_vault ? oci_kms_vault.k3s[0].id : ""
-      oci_region                        = var.region != null ? var.region : ""
-      external_secrets_chart_version    = var.external_secrets_chart_version
-      enable_dns01_challenge            = var.enable_dns01_challenge
-    })
+    content = join("\n", [
+      templatefile("${path.module}/files/server-vars.sh.tpl", {
+        k3s_version                       = local.k3s_version
+        k3s_subnet                        = var.k3s_subnet
+        k3s_token                         = var.enable_vault ? "" : random_password.k3s_token.result
+        k3s_url                           = local.k3s_internal_lb_ip
+        k3s_tls_san                       = local.k3s_internal_lb_ip
+        k3s_tls_san_public                = try(local.public_lb_ip[0], "")
+        expose_kubeapi                    = var.expose_kubeapi
+        compartment_ocid                  = var.compartment_ocid
+        availability_domain               = var.availability_domain
+        cluster_name                      = var.cluster_name
+        gitops_repo_url                   = var.gitops_repo_url
+        longhorn_ui_username              = var.longhorn_ui_username
+        longhorn_ui_password              = var.enable_vault ? "" : random_password.longhorn_ui_password.result
+        grafana_admin_password            = var.enable_vault ? "" : random_password.grafana_admin_password.result
+        vault_secret_id_k3s_token         = var.enable_vault ? oci_vault_secret.k3s_token[0].id : ""
+        vault_secret_id_longhorn_password = var.enable_vault ? oci_vault_secret.longhorn_ui_password[0].id : ""
+        vault_secret_id_grafana_password  = var.enable_vault ? oci_vault_secret.grafana_admin_password[0].id : ""
+        gateway_api_version               = var.gateway_api_version
+        certmanager_email_address         = var.certmanager_email_address
+        certmanager_chart_version         = var.certmanager_chart_version
+        argocd_chart_version              = var.argocd_chart_version
+        enable_external_dns               = var.enable_external_dns
+        cloudflare_api_token              = var.cloudflare_api_token != null ? var.cloudflare_api_token : ""
+        cloudflare_zone_id                = var.cloudflare_zone_id != null ? var.cloudflare_zone_id : ""
+        external_dns_domain_filter        = var.external_dns_domain_filter != null ? var.external_dns_domain_filter : ""
+        enable_external_secrets           = var.enable_external_secrets
+        vault_ocid                        = var.enable_vault ? oci_kms_vault.k3s[0].id : ""
+        oci_region                        = var.region != null ? var.region : ""
+        external_secrets_chart_version    = var.external_secrets_chart_version
+        enable_dns01_challenge            = var.enable_dns01_challenge
+        notification_topic_endpoint       = var.enable_notifications ? oci_ons_notification_topic.k3s_alerts[0].api_endpoint : ""
+        mysql_endpoint                    = var.enable_mysql ? "${oci_mysql_mysql_db_system.k3s[0].endpoints[0].hostname}:${oci_mysql_mysql_db_system.k3s[0].endpoints[0].port}" : ""
+        mysql_admin_username              = var.enable_mysql ? var.mysql_admin_username : ""
+        mysql_admin_password              = var.enable_mysql ? random_password.mysql_admin_password[0].result : ""
+      }),
+      file("${path.module}/files/lib/common.sh"),
+      file("${path.module}/files/lib/k3s-server.sh"),
+      file("${path.module}/files/lib/k3s-bootstrap.sh"),
+    ])
   }
 }
 
@@ -117,13 +110,17 @@ data "cloudinit_config" "k3s_worker" {
 
   part {
     content_type = "text/x-shellscript"
-    content = templatefile("${path.module}/files/k3s-install-agent.sh", {
-      k3s_version               = local.k3s_version
-      k3s_subnet                = var.k3s_subnet
-      k3s_token                 = var.enable_vault ? "" : random_password.k3s_token.result
-      k3s_url                   = local.k3s_internal_lb_ip
-      vault_secret_id_k3s_token = var.enable_vault ? oci_vault_secret.k3s_token[0].id : ""
-    })
+    content = join("\n", [
+      templatefile("${path.module}/files/agent-vars.sh.tpl", {
+        k3s_version               = local.k3s_version
+        k3s_subnet                = var.k3s_subnet
+        k3s_token                 = var.enable_vault ? "" : random_password.k3s_token.result
+        k3s_url                   = local.k3s_internal_lb_ip
+        vault_secret_id_k3s_token = var.enable_vault ? oci_vault_secret.k3s_token[0].id : ""
+      }),
+      file("${path.module}/files/lib/common.sh"),
+      file("${path.module}/files/lib/k3s-agent.sh"),
+    ])
   }
 }